split test into relevant files
