+++ src/public/services/__tests__/ExamplesService.test.ts
@@ -1,5 +1,8 @@
 import MockAxios from 'jest-mock-axios'
 
+import { DuplicateFormBody } from 'src/app/modules/form/admin-form/admin-form.types'
+import { IPopulatedUser, PublicForm, ResponseMode } from 'src/types'
+
 import * as ExamplesService from '../ExamplesService'
 
 jest.mock('axios', () => MockAxios)
@@ -88,4 +91,109 @@ describe('ExamplesService', () => {
       })
     })
   })
+
+  describe('useTemplate', () => {
+    it('should return template if POST request succeeds', async () => {
+      // Arrange
+      const MOCK_USER = {
+        _id: 'mock-user-id',
+      } as IPopulatedUser
+      const MOCK_FORM_ID = 'mock-form-id'
+      const expected = {
+        title: 'title',
+        lastModified: new Date(),
+        _id: MOCK_FORM_ID,
+        responseMode: ResponseMode.Email,
+        admin: MOCK_USER,
+      }
+      const MOCK_DUPLICATE_FORM_BODY = _generateDuplicateFormBody()
+
+      // Act
+      const actualPromise = ExamplesService.useTemplate(
+        MOCK_FORM_ID,
+        MOCK_DUPLICATE_FORM_BODY,
+      )
+      MockAxios.mockResponse({ data: expected })
+      const actual = await actualPromise
+
+      // Assert
+      expect(actual).toEqual(expected)
+      expect(MockAxios.post).toHaveBeenCalledWith(
+        `${ExamplesService.ADMIN_FORM_ENDPOINT}/${MOCK_FORM_ID}/adminform/copy`,
+        MOCK_DUPLICATE_FORM_BODY,
+      )
+    })
+
+    it('should reject with error message if POST request fails', async () => {
+      // Arrange
+      const expected = new Error('error')
+      const MOCK_FORM_ID = 'mock-form-id'
+      const MOCK_DUPLICATE_FORM_BODY = _generateDuplicateFormBody()
+
+      // Act
+      const actualPromise = ExamplesService.useTemplate(
+        MOCK_FORM_ID,
+        MOCK_DUPLICATE_FORM_BODY,
+      )
+      MockAxios.mockError(expected)
+
+      // Assert
+      await expect(actualPromise).rejects.toEqual(expected)
+      expect(MockAxios.post).toHaveBeenCalledWith(
+        `${ExamplesService.ADMIN_FORM_ENDPOINT}/${MOCK_FORM_ID}/adminform/copy`,
+        MOCK_DUPLICATE_FORM_BODY,
+      )
+    })
+  })
+
+  describe('queryTemplate', () => {
+    it('should return template if GET request succeeds', async () => {
+      // Arrange
+      const MOCK_USER = {
+        _id: 'mock-user-id',
+      } as IPopulatedUser
+      const MOCK_FORM_ID = 'mock-form-id'
+      const expected = ({
+        _id: MOCK_FORM_ID,
+        title: 'mock preview title',
+        admin: MOCK_USER,
+      } as unknown) as PublicForm
+
+      // Act
+      const actualPromise = ExamplesService.queryTemplate(MOCK_FORM_ID)
+      MockAxios.mockResponse({ data: expected })
+      const actual = await actualPromise
+
+      // Assert
+      expect(actual).toEqual(expected)
+      expect(MockAxios.get).toHaveBeenCalledWith(
+        `${ExamplesService.ADMIN_FORM_ENDPOINT}/${MOCK_FORM_ID}/adminform/template`,
+      )
+    })
+
+    it('should reject with error message if GET request fails', async () => {
+      // Arrange
+      const expected = new Error('error')
+      const MOCK_FORM_ID = 'mock-form-id'
+
+      // Act
+      const actualPromise = ExamplesService.queryTemplate(MOCK_FORM_ID)
+      MockAxios.mockError(expected)
+
+      // Assert
+      await expect(actualPromise).rejects.toEqual(expected)
+      expect(MockAxios.get).toHaveBeenCalledWith(
+        `${ExamplesService.ADMIN_FORM_ENDPOINT}/${MOCK_FORM_ID}/adminform/template`,
+      )
+    })
+  })
 })
+
+// Utils
+const _generateDuplicateFormBody = (): DuplicateFormBody => {
+  return {
+    title: 'title',
+    responseMode: ResponseMode.Email,
+    emails: 'test@example.com',
+  } as DuplicateFormBody
+}
